backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
motor>=3.3.0
beanie>=1.25.0
pytest-asyncio==0.21.1
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
httpx>=0.26.0
requests>=2.31.0
pytest>=8.0.0
python-dotenv>=1.0.1
redis>=5.0.1
playwright>=1.41.1
beautifulsoup4>=4.12.3
lxml>=5.1.0
openai>=1.10.0
jinja2>=3.1.3
aiofiles>=23.2.1
email-validator>=2.1.0.post1
APScheduler>=3.10.4
pytz>=2024.1
tenacity>=8.2.3
pypdf>=4.0.0
